refactor: break down GraphsScreen into smaller components
Extract 7 focused components from GraphsScreen.js (2,090 → 1,672 lines):
- BalanceHistoryCard: Line chart with balance history
- SpendingPredictionCard: Prediction stats with progress bar
- ExpenseSummaryCard: Mini expense summary with pie chart
- IncomeSummaryCard: Mini income summary with pie chart
- CustomLegend: Reusable legend for pie charts
- ExpensePieChart: Full expense pie chart with legend
- IncomePieChart: Full income pie chart with legend

This refactoring addresses issue #1 from the React Native best practices
plan by breaking down massive components into single-purpose, reusable
components. This improves:
- Maintainability: Easier to understand and modify
- Performance: More targeted re-renders
- Testability: Focused unit tests possible
- Code reuse: Logic extracted from monolithic screen

Part of issue #1: Component Size (CRITICAL)
